Rebased on ckan2
