+++ frictionless/package.py
@@ -294,20 +294,20 @@ class Package(Metadata):
         return storage.read_package()
 
     @staticmethod
-    def from_ckan(*, base_url, dataset_id, api_key=None):
+    def from_ckan(*, url, dataset, apikey=None):
         """Import package from CKAN
 
         Parameters:
-            base_url (str): (required) URL for CKAN instance (e.g: https://demo.ckan.org/ )
-            dataset_id (str): (required) ID or slug of dataset to fetch
-            api_key (str): (optional) Your CKAN API key
+            url (string): CKAN instance url e.g. "https://demo.ckan.org"
+            dataset (string): dataset id in CKAN e.g. "my-dataset"
+            apikey? (str): API key for CKAN e.g. "51912f57-a657-4caa-b2a7-0a1c16821f4b"
         """
         return Package.from_storage(
             system.create_storage(
-                "ckan_datastore",
-                base_url=base_url,
-                dataset_id=dataset_id,
-                api_key=api_key,
+                "ckan",
+                url=url,
+                dataset=dataset,
+                apikey=apikey,
             )
         )
 
@@ -471,21 +471,21 @@ class Package(Metadata):
         storage.write_package(self.to_copy(), force=force)
         return storage
 
-    def to_ckan(self, *, base_url, dataset_id=None, api_key=None, force=False):
+    def to_ckan(self, *, url, dataset, apikey=None, force=False):
         """Export package to CKAN
 
         Parameters:
-            base_url (str): (required) URL for CKAN instance (e.g: https://demo.ckan.org/ )
-            dataset_id (str): (optional) ID or slug of dataset this resource belongs to
-            api_key (str): (optional) Your CKAN API key
+            url (string): CKAN instance url e.g. "https://demo.ckan.org"
+            dataset (string): dataset id in CKAN e.g. "my-dataset"
+            apikey? (str): API key for CKAN e.g. "51912f57-a657-4caa-b2a7-0a1c16821f4b"
             force (bool): (optional) overwrite existing data
         """
         return self.to_storage(
             system.create_storage(
-                "ckan_datastore",
-                base_url=base_url,
-                dataset_id=dataset_id,
-                api_key=api_key,
+                "ckan",
+                url=url,
+                dataset=dataset,
+                apikey=apikey,
             ),
             force=force,
         )
